moved isadmin hook to alongside other components
