+++ components/client-fallback.tsx
@@ -7,7 +7,7 @@ import client from "@/tina/client";
 import { useQuery } from "@tanstack/react-query";
 import { notFound } from "next/navigation";
 import { useEffect } from "react";
-import { useIsAdminPage } from "./hooks/useIsAdmin";
+import { useIsAdminPage } from "../hooks/useIsAdmin";
 import { Container } from "./util/container";
 
 export interface ClientFallbackProps<T> {
